requirements.txt
groq>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Claude Agent SDK
claude-agent-sdk>=0.1.0
anyio>=4.0.0

# Web app
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.0
